:sparkles: Add a repost timeout
There might be a "race condition" when some other tool is trying to
set the presence, but the user has just selected a presence here. In
this case we repost, but only for a specific time. After that give
up on reposting and set the requested status to None, effectively
aborting the operation.

The timeout is currently fixed to 5 seconds.